requirements.txt
streamlit==1.56.0
langchain==0.2.17
langchain-community==0.2.17
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langchain-classic==0.2.17
zhipuai==2.1.5.20250825
faiss-cpu==1.13.2
python-dotenv==1.2.2
pypdf==5.4.0